[change] Split instance provider files
