--- Packages/UnitySimpleContainer
@@ -1,1 +1,0 @@
-com.witalosk.unity-simple-container
