+++ Packages/com.witalosk.unity-simple-container/Runtime/InstanceProviders.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 0e05188f32f34892880efa0bac7bbf09
+timeCreated: 1710499310
