+++ Packages/com.witalosk.unity-simple-container/Runtime/InstanceProviders/CollectionInstanceProvider.cs.meta
