+++ Packages/com.witalosk.unity-simple-container/Runtime/InstanceProviders/IInstanceProvider.cs
@@ -1,0 +1,7 @@
+namespace UnitySimpleContainer
+{
+    public interface IInstanceProvider
+    {
+        object GetInstance();
+    }
+}
