+++ Packages/com.witalosk.unity-simple-container/Runtime/InstanceProviders/IInstanceProvider.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: fbdd23a3cf6841f589debbc20081534a
+timeCreated: 1710499433
